+++ api/src/core/adapters/dbApi/kysely/kysely.utils.ts
@@ -25,3 +25,10 @@ export const castSql = <Db>(
     expression: SelectExpression<Db, keyof Db>,
     type: "int" | "text" | "bool" | "uuid"
 ): SelectExpression<Db, keyof Db> => sql`CAST(${expression} AS ${sql.raw(type)})` as any;
+
+export const isNotNull = <T>(value: T | null): value is T => value !== null;
+
+export const convertNullValuesToUndefined = <T extends Record<string, unknown>>(
+    obj: T
+): { [K in keyof T]: null extends T[K] ? Exclude<T[K], null> | undefined : T[K] } =>
+    Object.fromEntries(Object.entries(obj).map(([key, value]) => [key, value === null ? undefined : value])) as any;
